i don't know if EDF is working

+++ AlgorytmyDysku/src/Comparators/SortByDeadline.java
@@ -5,6 +5,6 @@ import java.util.Comparator;
 
 public class SortByDeadline implements Comparator<Request> {
     public int compare(Request a, Request b) {
-        return Integer.compare(a.getDeadline(), b.getDeadline());
+        return Double.compare(a.getDeadline(), b.getDeadline());
     }
 }
